+++ CHANGELOG.md
@@ -2,8 +2,8 @@
 
 ## 21.0-SNAPSHOT - unreleased
 
-### 🐞 Bug Fixes
-* Fix a regression to the configuration of Hazelcast distributed objects and hibernate caches.
+### ⚙️ Technical
+* Improvements to the ability to configure Hibernate Caches for 2nd-level and connection caches.
 
 ## 20.2.1 - 2024-07-09