+++ CHANGELOG.md
@@ -3,8 +3,9 @@
 ## 21.0-SNAPSHOT - unreleased
 
 ### ⚙️ Technical
-
-* Improvements to the ability to configure Hibernate Caches for 2nd-level and connection caches.
+*
+* Improvements to the ability to configure Hibernate 2nd-level caches. See `ClusterConfig` for more
+  information.
 
 ## 20.2.1 - 2024-07-09