+++ CHANGELOG.md
@@ -1,9 +1,9 @@
 # Changelog
 
-## 21.0-SNAPSHOT - unreleased
+## 20.3.0 - 2024-07-16
 
 ### ⚙️ Technical
-*
+
 * Improvements to the ability to configure Hibernate 2nd-level caches. See `ClusterConfig` for more
   information.